+++ packages/dart_edge_core/lib/src/annotations.dart
@@ -1,1 +1,4 @@
-export 'annotations/from_schema.dart';
+export 'package:json_schema/json_schema.dart' show FromSchema, SchemaRefModel;
+
+export 'annotations/from_http_schema.dart';
+export 'annotations/from_multipart_schema.dart';