+++ packages/dart_edge_core/lib/src/http/json_schema_registry.dart
@@ -1,0 +1,1 @@
+export 'package:json_schema/json_schema.dart' show JsonSchemaRegistry;